refactor: Move utility functions
