+++ packages/scanner/src/util/eventAggregator.ts
