+++ packages/scanner/src/util/isAncestorPath.ts
